refactor: split HelpMenu into separate module
